% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_frac_abund.R
\name{add_frac_abund}
\alias{add_frac_abund}
\title{add_frac_abund}
\usage{
add_frac_abund(isotope_list)
}
\arguments{
\item{isotope_list}{a list from the enviPat package of matrices, indexed by
chemical formula, containing isotopes, isotopic abundances, and atom
numbers}
}
\description{
add_frac_abund
}
\examples{
add_frac_abund(subset_SCCP_isotope_list)
}
