% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_conc_CF.R
\name{adjust_conc_CF}
\alias{adjust_conc_CF}
\title{adjust_conc_CF}
\usage{
adjust_conc_CF(MSdata, model = "RF_model")
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data}

\item{model}{The computed model used to predict the mixture
concentration ("RF_model" or "calib_model")}
}
\value{
an augmented tibble containing the solid concentration
}
\description{
adjust_conc_CF
}
