% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_SCCP_congeners.R
\name{aggregate_SCCP_congeners}
\alias{aggregate_SCCP_congeners}
\title{aggregate_SCCP_congeners}
\usage{
aggregate_SCCP_congeners(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data, normalized by the internal standard}
}
\value{
a tibble containing the aggregated SCCPs
}
\description{
aggregate_SCCP_congeners
}
