% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_SCCP_rel_peak_areas.R
\name{aggregate_SCCP_rel_peak_areas}
\alias{aggregate_SCCP_rel_peak_areas}
\title{aggregate_SCCP_rel_peak_areas}
\usage{
aggregate_SCCP_rel_peak_areas(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data}
}
\value{
a tibble containing the aggregated SCCPs
}
\description{
aggregate_SCCP_rel_peak_areas
}
