% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_calib_model_zero_intercept.R
\name{compute_calib_model_zero_intercept}
\alias{compute_calib_model_zero_intercept}
\title{compute_calib_model_zero_intercept}
\usage{
compute_calib_model_zero_intercept(MSdata, model = "quadratic")
}
\arguments{
\item{MSdata}{a tibble containing the response factors and wt% Cl of a set of
SCCP technical mixture standards}

\item{model}{a character string indicating whether a "linear", "quadratic" or
"exponential" model was used}
}
\description{
compute_calib_model_zero_intercept
}
