% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Cl_perc.R
\name{compute_Cl_perc}
\alias{compute_Cl_perc}
\title{compute_Cl_perc}
\usage{
compute_Cl_perc(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder for the SCCP
congeners}
}
\value{
a tibble containing MSdata with chlorine data
}
\description{
compute_Cl_perc
}
