% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_exper_isotop_abund.R
\name{compute_exper_isotop_abund}
\alias{compute_exper_isotop_abund}
\title{compute_exper_isotop_abund}
\usage{
compute_exper_isotop_abund(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}
}
\value{
an augmented tibble with relative isotopic abundance for each ion
}
\description{
compute_exper_isotop_abund
}
