% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RF_model.R
\name{compute_RF_model}
\alias{compute_RF_model}
\title{compute_RF_model}
\usage{
compute_RF_model(MSdata, model = "quadratic")
}
\arguments{
\item{MSdata}{a tibble containing the response factors and wt% Cl of a set of
SCCP technical mixture standards}

\item{model}{a character string in c("linear", "quadratic", "exponential")}
}
\description{
compute_RF_model
}
