% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RF_tech_mix.R
\name{compute_RF_tech_mix}
\alias{compute_RF_tech_mix}
\title{compute_RF_tech_mix}
\usage{
compute_RF_tech_mix(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data}
}
\value{
an augmented tibble including computed RF data
}
\description{
compute_RF_tech_mix
}
