% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_surrogate_calib.R
\name{compute_surrogate_calib}
\alias{compute_surrogate_calib}
\title{compute_surrogate_calib}
\usage{
compute_surrogate_calib(MSdata, model = "quadratic")
}
\arguments{
\item{MSdata}{a tibble containing the surrogate calibration standards}

\item{model}{a character string indicating whether a "linear", "quadratic" or
"exponential" model was used}
}
\description{
compute_surrogate_calib
}
