% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_sample_names.R
\name{correct_sample_names}
\alias{correct_sample_names}
\title{correct_sample_names}
\usage{
correct_sample_names(MSdata)
}
\arguments{
\item{MSdata}{a tibble of MS data with incorrect sample names}
}
\value{
a corrected tibble
}
\description{
correct_sample_names
}
