% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_col_names.R
\name{customize_col_names}
\alias{customize_col_names}
\title{customize_col_names}
\usage{
customize_col_names(MSdata = total_SCCPs)
}
\arguments{
\item{MSdata}{a list of tibbles containing MSdata for the SCCP congeners
(written for total_SCCPs)}
}
\value{
a list of tibbles with customized column names
}
\description{
customize_col_names
}
