% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conc_zero_intercept.R
\name{estimate_conc_zero_intercept}
\alias{estimate_conc_zero_intercept}
\title{estimate_conc_zero_intercept}
\usage{
estimate_conc_zero_intercept(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data}
}
\value{
an augmented tibble containing the predicted concentration
}
\description{
estimate_conc_zero_intercept
}
