% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_solid_conc.R
\name{estimate_solid_conc}
\alias{estimate_solid_conc}
\title{estimate_solid_conc}
\usage{
estimate_solid_conc(MSdata, model = "RF_model")
}
\arguments{
\item{MSdata}{a tibble containing SCCP congener data}

\item{model}{The computed model used to predict the mixture
concentration ("RF_model" or "calib_model")}
}
\value{
an augmented tibble containing the solid concentration
}
\description{
estimate_solid_conc
}
