% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_surrogate_conc.R
\name{estimate_surrogate_conc}
\alias{estimate_surrogate_conc}
\title{estimate_surrogate_conc}
\usage{
estimate_surrogate_conc(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing surrogate data}
}
\value{
an augmented tibble containing the predicted concentration
}
\description{
estimate_surrogate_conc
}
