% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_CP_isotope_df.R
\name{gen_CP_isotope_df}
\alias{gen_CP_isotope_df}
\title{gen_CP_isotope_df}
\usage{
gen_CP_isotope_df(isotope_list = SCCP_isotope_list, top_n_num = 3)
}
\arguments{
\item{isotope_list}{a list of matrices, indexed by chemical formula,
containing isotopes with absolute and fractional relative abundances for
each chemical formula}

\item{top_n_num}{the desired numer of isotopes for each chemical formula}
}
\value{
a dataframe containing CP formulas, ions, and isotopic abundances
}
\description{
gen_CP_isotope_df
}
\examples{
gen_CP_isotope_df(subset_enhanced_SCCP_isotope_list, top_n_num = 4)
}
