% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_list.R
\name{get_element_list}
\alias{get_element_list}
\title{get_element_list}
\usage{
get_element_list(elements_df)
}
\arguments{
\item{elements_df}{A dataframe with chemical elements as column names.}
}
\value{
A vector of chemical element names.
}
\description{
Receives a named dataframe containing columns of chemical element counts and
returns a vector of element names in Hill System Order (C, H, alphabetical).
}
\examples{
test_df <- data.frame(O = 3, F = 5, H = 1, C = 2, S = 1)
get_element_list(test_df)
}
