% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enhanced_isotope_list.R
\name{get_enhanced_isotope_list}
\alias{get_enhanced_isotope_list}
\title{get_enhanced_isotope_list}
\usage{
get_enhanced_isotope_list(chemformulas)
}
\arguments{
\item{chemformulas}{a character string or vector of characters strings
representing chemical formulas}
}
\value{
a list of matrices, indexed by chemical formula, containing isotopes,
  relative isotopic abundances, and fractional relative isotopic abundances
  for each chemical formula
}
\description{
get_enhanced_isotope_list
}
\examples{
formula <- "C17H22BrF13"
get_enhanced_isotope_list(formula)

subset_SCCP_formulas <- SCCP_formulas[1:3]
get_enhanced_isotope_list(subset_SCCP_formulas)
}
