% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inject_stds.R
\name{get_inject_stds}
\alias{get_inject_stds}
\title{get_inject_stds}
\usage{
get_inject_stds(MSdata, std = "13C-HCB")
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}

\item{std}{a character string containing the injection standard name, as
specified in the "compound" column. The default is the injection standard
used during method development.}
}
\value{
a tibble containing the injection standard data
}
\description{
get_inject_stds
}
