% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isotope_list.R
\name{get_isotope_list}
\alias{get_isotope_list}
\title{get_isotope_list}
\usage{
get_isotope_list(chemformulas)
}
\arguments{
\item{chemformulas}{a character string or vector of characters strings
representing chemical formulas}
}
\value{
a list of matrices from enviPat, indexed by chemical formula,
  containing isotopes and abundances for each chemical formula
}
\description{
get_isotope_list
}
\examples{
formula <- "C17H22BrF13"
get_isotope_list(formula)

subset_SCCP_formulas <- SCCP_formulas[1:3]
get_isotope_list(subset_SCCP_formulas)
}
