% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isotope_mass.R
\name{get_isotope_mass}
\alias{get_isotope_mass}
\title{get_isotope_mass}
\usage{
get_isotope_mass(iso)
}
\arguments{
\item{iso}{character string representing elemental isotope}
}
\value{
numeric representing the elemental isotope's mass
}
\description{
Returns mass for specific elemental isotope (e.g., 1H or 35Cl)
}
\examples{
get_isotope_mass("35Cl")
}
