% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_inj_stds.R
\name{get_new_inj_stds}
\alias{get_new_inj_stds}
\title{get_new_inj_stds}
\usage{
get_new_inj_stds(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}
}
\value{
a tibble containing the surrogate data
}
\description{
get_new_inj_stds
}
