% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quant_ions.R
\name{get_quant_ions}
\alias{get_quant_ions}
\title{get_surrogate_ions}
\usage{
get_quant_ions(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}
}
\value{
a tibble containing only the quantitation ions, which are labeled as "peak_label" = 'T1'
}
\description{
get_surrogate_ions
}
