% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SCCP_congeners.R
\name{get_SCCP_congeners}
\alias{get_SCCP_congeners}
\title{get_SCCP_congeners}
\usage{
get_SCCP_congeners(MSdata, inject_std = "13C-HCB",
  surrogate = "13C-PCN27")
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}

\item{inject_std}{a character string containing the injection standard name,
as specified in the "compound" column. The default is the injection
standard used during method development.}

\item{surrogate}{a character string containing the surrogate standard name,
as specified in the "compound" column. The default is the surrogate
standard used during method development.}
}
\description{
get_SCCP_congeners
}
