% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surrog_calib.R
\name{get_surrog_calib}
\alias{get_surrog_calib}
\title{get_surrog_calib}
\usage{
get_surrog_calib(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing the surrogate data from the calib standards}
}
\value{
a tibble containing the surrogate calib standards
}
\description{
get_surrog_calib
}
