% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surrogate_quant_ions.R
\name{get_surrogate_quant_ions}
\alias{get_surrogate_quant_ions}
\title{get_surrogate_quant_ions}
\usage{
get_surrogate_quant_ions(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing the surrogate data}
}
\value{
a tibble containing the surrogate quantification ions
}
\description{
get_surrogate_quant_ions
}
