% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surrogates.R
\name{get_surrogates}
\alias{get_surrogates}
\title{get_surrogates}
\usage{
get_surrogates(MSdata, surrogate = "13C-PCN27")
}
\arguments{
\item{MSdata}{a tibble containing MSdata from TraceFinder}

\item{surrogate}{a character string containing the surrogate standard name,
as specified in the "compound" column. The default is the surrogate
standard used during method development.}
}
\value{
a tibble containing the surrogate data
}
\description{
get_surrogates
}
