% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TFdata.R
\name{get_TFdata}
\alias{get_TFdata}
\title{get_TFdata}
\usage{
get_TFdata(filename)
}
\arguments{
\item{filename}{A character string containing the name of the Excel file that
contains the TraceFinder MS data.}
}
\value{
a tibble containing MS data for the desired technical mixture
 standards.
}
\description{
This function reads an Excel file exported by TraceFinder (one sample per
worksheet), cleans up the filenames, drops empty columns, converts numbers
occuring as strings, and returns a subset of the data when abbrev = TRUE.
}
