% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{Isotopes}
\format{A data frame with 295 rows and 5 variables: \describe{
  \item{element}{Common symbol of chemical element, a character string}
  \item{isotope}{Isotope symbol derived from the stable isotope's unit mass
  and common symbol, a character string} \item{mass}{accurate isotopic mass,
  in decimal unified atomic mass units (u), a numeric}
  \item{abundance}{relative isotopic abundance, in decimal rather than
  percent notation, a numeric} \item{ratioC}{Maximum number of atoms for one
  C-atom in a molecule, used in enviPat, an integer} ... }}
\source{
adapted from the enviPat package for use with enviPat::isopattern and
  for internal calculations
}
\usage{
isotopes
}
\description{
A dataset containing the stable isotopes of 84 chemical elements with their
accurate masses and relative isotopic abundances
}
\examples{
data(isotopes)
}
\keyword{datasets}
