% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot_calib}
\alias{plot_calib}
\title{plot_calib}
\usage{
plot_calib(MSdata, model, type = "exponential")
}
\arguments{
\item{MSdata}{a tibble containing response and concentration of a technical
mixture or surrogate standard}

\item{model}{a list containing output from the lm function}

\item{type}{a character string indicating whether a "linear", "quadratic" or
"exponential" model was used}
}
\value{
plot as side effect
}
\description{
plot_calib
}
