% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isotopic_pattern.R
\name{plot_isotopic_pattern}
\alias{plot_isotopic_pattern}
\title{plot_isotopic_pattern}
\usage{
plot_isotopic_pattern(chem_formula)
}
\arguments{
\item{chem_formula}{a character string containing a molecular formula
reference to the enhanced SCCP isotope list}
}
\value{
produces a plot of isotopic abundances
}
\description{
plot_isotopic_pattern
}
\examples{
chem_formula <- "C10H15Cl7"
plot_isotopic_pattern(chem_formula)
}
