% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RF_model.R
\name{plot_RF_model}
\alias{plot_RF_model}
\title{plot_RF_model}
\usage{
plot_RF_model(MSdata, regr, type = "linear", colorBy = "TMstd")
}
\arguments{
\item{MSdata}{a tibble containing response factor and \%Cl data for a range of
SCCP technical mixture standards}

\item{regr}{a list containing output from the lm function}

\item{type}{a character string indicating whether a "linear" or "quadratic"
model was used}

\item{colorBy}{a character string indicating whether to color by "TMstd"
(designated by its nominal \% Cl) or "concentration".}
}
\value{
plot as side effect
}
\description{
plot_RF_model
}
