% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surrogate_calib.R
\name{plot_surrogate_calib}
\alias{plot_surrogate_calib}
\title{plot_surrogate_calib}
\usage{
plot_surrogate_calib(MSdata, model = surrogate_calib_model,
  type = "quadratic")
}
\arguments{
\item{MSdata}{a tibble containing the surrogate calibration standards}

\item{model}{a list containing output from the lm function}

\item{type}{a character string indicating whether a "linear", "quadratic" or
"exponential" model was used}
}
\value{
plot as side effect
}
\description{
plot_surrogate_calib
}
