% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_MSdata_worksheet.R
\name{save_MSdata_worksheet}
\alias{save_MSdata_worksheet}
\title{save_MSdata_worksheet}
\usage{
save_MSdata_worksheet(worksheet, directory = choose.dir(), desired_name)
}
\arguments{
\item{worksheet}{a tibble to be saved as csv}

\item{directory}{a full directory path (default is interactive directory
selection)}

\item{desired_name}{a string containing the desired filename (".csv" will be
appended)}
}
\description{
save_MSdata_worksheet
}
