% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCP_formulas.R
\docType{data}
\name{SCCP_formulas}
\alias{SCCP_formulas}
\title{SCCP_formulas}
\format{A vector of 24 character strings}
\source{
gen_CP_formulas package
}
\usage{
SCCP_formulas
}
\description{
A dataset containing the chemical formulas of short-chain chlorinated
paraffins with carbon numbers ranging from 10 to 13 and chlorine numbers
ranging from 5 to 10
}
\examples{
data(SCCP_formulas)
}
\keyword{datasets}
