% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCP_isotope_df.R
\docType{data}
\name{SCCP_isotope_df}
\alias{SCCP_isotope_df}
\title{SCCP_isotope_df}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72 rows and 14 columns.}
\source{
gen_enhanced_isotope_list package
}
\usage{
SCCP_isotope_df
}
\description{
A dataframe generated by the default settings in gen_CP_isotope_df. It
contains data for the three (3) most abundant isotopes of Short-chain
chlorinated paraffins (SCCPs) with carbon numbers ranging from 10 to 13 and
chlorine numbers ranging from 5 to 10. It is used by default for most SCCP
processing.
}
\examples{
data(SCCP_isotope_df)
}
\keyword{datasets}
