% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCP_isotope_list.R
\docType{data}
\name{SCCP_isotope_list}
\alias{SCCP_isotope_list}
\title{SCCP_isotope_list}
\format{An object of class \code{list} of length 24.}
\source{
gen_enhanced_isotope_list package
}
\usage{
SCCP_isotope_list
}
\description{
A dataset containing isotopic distributions for the chemical formulas of
short-chain chlorinated paraffins with carbon numbers ranging from 10 to 13
and chlorine numbers ranging from 5 to 10, created with enviPat::isopattern
and internal functions for use in processing SCCP data.
}
\examples{
data(SCCP_isotope_list)
}
\keyword{datasets}
