% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_MSdata.R
\name{split_MSdata}
\alias{split_MSdata}
\title{split_MSdata}
\usage{
split_MSdata(MSdata)
}
\arguments{
\item{MSdata}{a tibble containing the imported TraceFinder data, and
containing a column which specifies the user-defined sample type}
}
\value{
a list of tibbles, separated by sample type
}
\description{
split_MSdata
}
