% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sample_id.R
\name{split_sample_id}
\alias{split_sample_id}
\title{split_sample_id}
\usage{
split_sample_id(MSdata)
}
\arguments{
\item{MSdata}{a list of tibbles containing MSdata for the SCCP congeners}
}
\value{
a list of tibbles with customized column names from the "sample_id" column
}
\description{
split_sample_id
}
