% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{transpose_df}
\usage{
transpose_df(df)
}
\arguments{
\item{df}{a tibble or dataframe to transpose (e.g., rows to columns)}
}
\value{
a transposed tibble or dataframe
}
\description{
transpose_df
}
